refactor: 💡 Migrate SQForm to TS
Migrate SQForm.js to TypeScript

✅ Closes: #346
